auth-app/backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
pymongo==4.5.0
bcrypt==4.0.1
PyJWT==2.8.0
python-dotenv==1.0.0
implicit==0.7.2
scipy==1.11.3
numpy==1.26.0
pandas==2.1.1
scikit-learn==1.3.1
langdetect==1.0.9
redis==5.0.1
gunicorn==21.2.0
gevent==23.9.1
faiss-cpu==1.7.4
joblib==1.3.2
psutil==5.9.6
openai>=1.0.0
